remove cutover-era postgres-template scripts and committed auth JSONs
