--- e2e/.auth/A1001.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "30w9KP0lOo18OvbAuvlQFNzHqAL2w82D.i%2BY%2BSPkFORr6xH5PkoES6VvcAU9p3X9CY67rHdZwSTE%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917244.298966,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
