--- e2e/.auth/A1101.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "5VnNGiRdWSKMwNdDScUc5kBqOyx0gSus.lpMWUL2foPjAIYybWvlKV%2FW5akwhJoPvV8RfPTy5xlk%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917244.18855,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
