--- e2e/.auth/A1102.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "exPUg3fF80vXTetRPTB8FMHklNzWze58.Ir99hcrkezXBXC8%2BH0%2FDJhEzR%2Fiep9V%2FCVi1OLrpnDM%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917244.293684,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
