--- e2e/.auth/A1201.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "GNPeKxbqvSs85whiN0FIjYG5cH9ekT7z.kRf05BSRmL8Lh1b%2BziGyR5C8NLx6SsRq9QDCRUiM3%2F4%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917244.457887,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
