--- e2e/.auth/A1302.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "36xyQgU1hkdpmfCF4wh9xDT2cQW8pLPa.JuXWYPmWFOGLVuDcVAZe1bjlZdhX3oV%2FWKn%2BldIRU5o%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917244.29423,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
