--- e2e/.auth/B1101.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "NSeFLwwA2RrcfoIfkQt12nz7knIfY7SO.fhj0S%2B1S7kobd7eH2MNlNx8l6CKGh0FrpBuHi06WmlQ%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917245.231917,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
