--- e2e/.auth/B1201.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "DcNfW07bsZmFeIqlXJZnyCfIdrVMDPEa.VzEk4wHwvooUaVIMJj%2Blf0OAcS8Aq2cfpdiJcRttFEk%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917245.300542,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
