--- e2e/.auth/B1202.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "BALzIah3ySEkiPUKok8PodolKri6LhKi.h9K9ghXU85%2B1oljjFj%2F9lzgcMS5vdqwtrdF1057axm0%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917245.309866,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
