--- e2e/.auth/B1301.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "aJUz0pnSuS9WEMRrdWRKR7PFvi2wxiDG.z5Aa3xSRY%2BwKa%2FXuj8MXIBr0gxoFeAv3LCq836GEg9s%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917245.309497,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
