--- e2e/.auth/B1302.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "Qm62KlKYCNbR4C75nAvx3lLVm1BuVFDU.kfi1jQK9NO4xn%2FYR3Oly8cLT%2F7%2F6e5xO5My9YQAuCtA%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917245.373397,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
