--- e2e/.auth/C1001.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "PDYOklllq9MZniUzY7jHeLBlKT9CBtrE.Et1Ci2HfFuHWgdMn%2FuWe9NWeKwFEkrdyTxeubDS5BWc%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917246.352243,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
