--- e2e/.auth/C1002.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "x852pjqhspeJXDYawBnXGnmTtn4D6frq.ZqqtOaZHu1LIvXnghTVH864Z92xXNQFJbvPiRuxgJ3g%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917246.281393,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
