--- e2e/.auth/C1101.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "gIJueFE8WGV41hccl1pOpwX32FOjWLkc.lTMSca8B6yKnVeMn2X5T6nfWoIDkIt6H7FMI%2FDQ4Ntw%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917246.347965,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
