--- e2e/.auth/C1102.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "P7laPmFAxBzuyIGV6oJDAT1cV3jxdEpf.es79pEFPQfzZOdLEUd5Y832Sxgn95YtopwMeXZtCjnE%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917246.283902,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
