--- e2e/.auth/C1302.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "tlMUPUTSODuXmcr1jaXPITJ1Iib9NgtJ.BsxwiVBfCn9ivbeimdeaoOeHoz4X7DO93iHm7%2Bvktpw%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917246.285338,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
