--- e2e/.auth/D1002.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "2z63BNLkwicWcgt7WJce61MMZaVEqIgn.kV0ACTdGAMI%2F0VQxq4yU6TEjMwvch4VG83qmZ9dgMDo%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917247.256085,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
