--- e2e/.auth/D1101.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "ZX1TzkitByOK9CvXVRDGAjDCVVbyuKEq.qVNSLrBuXt8adq9ZCF%2BFQNPRX6pGoDNGSZBGyYsPlWs%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917247.199064,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
