--- e2e/.auth/D1102.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "F8aipB6FH5Xa98Eq20CVKZBhzj11nstO.0MFMFc%2F0WyuBehLF8fqpM566DVkkRO2LDGPOACc3BDw%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917247.307901,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
