--- e2e/.auth/D1201.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "3Hw5kSE5gvdNYNJPcBCzPir4lr7c0k2w.lecipfPd%2BNrMuBQK%2F6BMAj7xIbpDg0EdfM6Dm8EMB10%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917247.307544,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
