--- e2e/.auth/D1202.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "BdImMxI14sFPfaylnfSaZAI06GWnRXoE.j2GZPvZ6oOE5p0eT3IIMQrcYsfMCxtUdmfIHmQSjUcQ%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917247.379895,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
