--- e2e/.auth/D1301.json
@@ -1,15 +1,0 @@
-{
-  "cookies": [
-    {
-      "name": "better-auth.session_token",
-      "value": "HqefQagEvQlR1vhJotAaXwNDrgStaIIU.e%2Fjikjvu3W%2F8dMN4xw0N3DEfjbm1S2Z0V36LSGqkEjo%3D",
-      "domain": "localhost",
-      "path": "/",
-      "expires": 1777917247.328837,
-      "httpOnly": true,
-      "secure": false,
-      "sameSite": "Lax"
-    }
-  ],
-  "origins": []
-}
